useClient filter hook was implemented
