+++ src/app/ui/ClientFilter/ClientFilter.tsx
@@ -1,12 +1,9 @@
-import { useCallback, useState } from 'react';
 import DropDown from '../DropDown/DropDown';
 import classNames from 'classnames';
-import { translateGenre } from '@/app/shared/translateGenre';
 import { Genre } from '@/store/features/filterSlice';
 import { Cinema } from '@/app/types';
-import useDropDown from '@/app/hooks/useDropDown';
 import styles from './clietnFilter.module.css';
-import useAppContext from '@/app/hooks/useAppContext';
+import useClientFilter from './useClientFilter';
 
 interface ClientFilterProps {
   title: string;
@@ -17,22 +14,14 @@ interface ClientFilterProps {
 
 export default function ClientFilter (props: ClientFilterProps) {
   const { title, placeholder, className, variants } = props;
-  const { isOpen, toggleOpen, closeDropDown } = useDropDown(styles.wrapper);
-  
-  const { genrePlaceholder, setGenrePlaceholder} = useAppContext();
-  const [curValue, setCurValue] = useState<string>(genrePlaceholder ||placeholder);
 
-  const setCurrentName = useCallback((value: Genre | Cinema) => {
-    if (typeof value !== 'object') {
-      if (value === Genre.DEFAULT) {
-        setCurValue(title);
-        setGenrePlaceholder?.(title);
-      } else {
-        setCurValue(translateGenre(value));
-        setGenrePlaceholder?.(translateGenre(value));
-      }
-    }
-  }, [title, setGenrePlaceholder]);
+  const { 
+    isOpen, 
+    toggleOpen, 
+    curValue, 
+    closeDropDown, 
+    setCurrentName 
+  } = useClientFilter(styles.wrapper, placeholder, title);
 
   return (
     <div className={classNames(className, styles.wrapper)}>
